LPS-124613 Use new method to check index existence before removing it

+++ modules/apps/asset/asset-service/src/main/java/com/liferay/asset/internal/upgrade/v2_0_1/UpgradeAssetEntryUsage.java
@@ -23,7 +23,9 @@ public class UpgradeAssetEntryUsage extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		runSQL("drop index IX_71A0231C on AssetEntryUsage");
+		if (hasIndex("AssetEntryUsage", "IX_71A0231C")) {
+			runSQL("drop index IX_71A0231C on AssetEntryUsage");
+		}
 	}
 
 }
